Undo unrelated test changes
